Added a way to dynamically spawn the player character based off the class that they chose. This character will be created by the HeroFactory and inherit from the Hero class. I created a way for the camera to stay loaded by implementing it into the managerScene. I created the Thief, Mage, and modified the Knight (now Warrior classes). I added prefabs for ease of placement in scenes. I created custom animations controllers for each of the three classes. I still need to modify the rest of the scenes to be able to accept the new camera and player character. I tied the character selector from the start menu into the spawning of the player character. I implemented the customPlayerFlip and CustomPlayerGraphic classes.

+++ Assets/Scripts/Model/DungeonCharacter.cs
@@ -1,3 +1,5 @@
+using Model.JSON;
+
 namespace Model
 {
     public abstract class DungeonCharacter
@@ -6,20 +8,20 @@ namespace Model
         private int hitPoints;
         private int damageMin;
         private int damageMax;
-        private int attackSpeed;
-        private int moveSpeed;
+        private float attackSpeed;
+        private float moveSpeed;
         private float chanceToCrit;
 
         public string Name { get => name; set => name = value; }
         public int HitPoints { get => hitPoints; protected set => hitPoints = value; }
         public int DamageMin { get => damageMin; protected set => damageMin = value; }
         public int DamageMax { get => damageMax; protected set => damageMax = value; }
-        public int AttackSpeed { get => attackSpeed; protected set => attackSpeed = value; }
-        public int MoveSpeed { get => moveSpeed; protected set => moveSpeed = value; }
+        public float AttackSpeed { get => attackSpeed; protected set => attackSpeed = value; }
+        public float MoveSpeed { get => moveSpeed; protected set => moveSpeed = value; }
         public float ChanceToCrit { get => chanceToCrit; protected set => chanceToCrit = value; }
 
         protected DungeonCharacter(string name, int hitPoints, int damageMin, int damageMax, 
-            int attackSpeed, int moveSpeed, float chanceToCrit)
+            float attackSpeed, float moveSpeed, float chanceToCrit)
         {
             this.name = name;
             this.hitPoints = hitPoints;
